requirements.txt
aiolimiter==1.1.0
anyio==4.3.0
APScheduler==3.10.4
cachetools==5.3.2
certifi==2024.2.2
dnspython==2.6.1
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
pyaes==1.6.1
pymongo==4.6.2
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
python-telegram-bot==20.8
pytz==2024.1
six==1.16.0
sniffio==1.3.1
TgCrypto==1.2.5
tornado==6.4
tzdata==2024.1
tzlocal==5.2
